Simplify tote search... just looking for rectangular yellow blod now.
Incusion tester is upgraded to return the numbe of matched points. We are using
it to detect the yellow tote shape. Only 7 of the 9 tested points must match.
This is for when the test points land on one of the non-yellow L's, and/or
for when 1 or 2 points land outside because the tote is skewed.

Aspect ration, perimeter area checks are all disabled for now.

This algorithm seems to be working quite well on the frc sample images. We'll
adapt as needed during live testing.